requirements.txt
django
djangorestframework
scrapy
dj_database_url
uwsgi
